+++ src/cpr/cpr_parameter_pack.cxx
@@ -57,15 +57,15 @@ void piecewise_parameter_pack::write_to_file(std::ofstream& file) const{
   this->parameter_pack::write_to_file(file);
   file << this->num_dimensions << "\n";
   file << this->num_knots << "\n";
-  for (int i=0; i<num_partitions_per_dimension.size(); i++){
+  for (size_t i=0; i<num_partitions_per_dimension.size(); i++){
     if (i>0) file << ",";
     file << num_partitions_per_dimension[i];
   } file << "\n";
-  for (int i=0; i<knot_index_offsets.size(); i++){
+  for (size_t i=0; i<knot_index_offsets.size(); i++){
     if (i>0) file << ",";
     file << knot_index_offsets[i];
   } file << "\n";
-  for (int i=0; i<knot_positions.size(); i++){
+  for (size_t i=0; i<knot_positions.size(); i++){
     if (i>0) file << ",";
     file << knot_positions[i];
   } file << "\n";
@@ -80,17 +80,17 @@ void piecewise_parameter_pack::read_from_file(std::ifstream& file){
   this->knot_index_offsets.resize(this->num_dimensions);
   this->knot_positions.resize(this->num_knots);
   std::string temp;
-  for (int i=0; i<num_partitions_per_dimension.size(); i++){
+  for (size_t i=0; i<num_partitions_per_dimension.size(); i++){
     if (i==(num_partitions_per_dimension.size()-1)) getline(file,temp,'\n');
     else getline(file,temp,',');
     this->num_partitions_per_dimension[i] = std::stoi(temp);
   }
-  for (int i=0; i<knot_index_offsets.size(); i++){
+  for (size_t i=0; i<knot_index_offsets.size(); i++){
     if (i==(knot_index_offsets.size()-1)) getline(file,temp,'\n');
     else getline(file,temp,',');
     this->knot_index_offsets[i] = std::stoi(temp);
   }
-  for (int i=0; i<knot_positions.size(); i++){
+  for (size_t i=0; i<knot_positions.size(); i++){
     if (i==(knot_positions.size()-1)) getline(file,temp,'\n');
     else getline(file,temp,',');
     this->knot_positions[i] = std::stod(temp);
@@ -103,7 +103,7 @@ tensor_parameter_pack::tensor_parameter_pack() : piecewise_parameter_pack(){
 }
 
 tensor_parameter_pack::tensor_parameter_pack(const tensor_parameter_pack& rhs) : piecewise_parameter_pack(rhs){
-  int nelements_rhs = 0;
+  size_t nelements_rhs = 0;
   for (const auto& it : rhs.num_partitions_per_dimension) nelements_rhs *= it;
   assert(rhs.num_tensor_elements == nelements_rhs);
   this->tensor_elements = new double[nelements_rhs];
@@ -127,7 +127,7 @@ void tensor_parameter_pack::set(const parameter_pack& rhs){
   this->piecewise_parameter_pack::set(rhs);
   const tensor_parameter_pack& rhs_derived = dynamic_cast<const tensor_parameter_pack&>(rhs);
   if (this->tensor_elements != nullptr) delete[] this->tensor_elements;
-  int nelements_rhs = 0;
+  size_t nelements_rhs = 0;
   for (const auto& it : rhs_derived.num_partitions_per_dimension) nelements_rhs *= it;
   assert(nelements_rhs == rhs_derived.num_tensor_elements);
   this->tensor_elements = new double[nelements_rhs];
@@ -140,7 +140,7 @@ void tensor_parameter_pack::write_to_file(std::ofstream& file) const{
   this->piecewise_parameter_pack::write_to_file(file);
   if (this->tensor_elements == nullptr) return;
   file << this->num_tensor_elements << "\n";
-  for (int64_t i=0; i<this->num_tensor_elements; i++){
+  for (size_t i=0; i<this->num_tensor_elements; i++){
     if (i>0) file << ",";
     file << this->tensor_elements[i];
   } file << "\n";
@@ -152,7 +152,7 @@ void tensor_parameter_pack::read_from_file(std::ifstream& file){
   file >> this->num_tensor_elements;
   if (this->tensor_elements == nullptr) this->tensor_elements = new double[this->num_tensor_elements];
   std::string temp;
-  for (int i=0; i<this->num_tensor_elements; i++){
+  for (size_t i=0; i<this->num_tensor_elements; i++){
     if (i==(this->num_tensor_elements-1)) getline(file,temp,'\n');
     else getline(file,temp,',');
     this->tensor_elements[i] = std::stod(temp);
@@ -167,7 +167,7 @@ cpr_parameter_pack::cpr_parameter_pack() : piecewise_parameter_pack(){
 }
 
 cpr_parameter_pack::cpr_parameter_pack(const cpr_parameter_pack& rhs) : piecewise_parameter_pack(rhs){
-  int nelements_rhs = 0;
+  size_t nelements_rhs = 0;
   for (const auto& it : rhs.num_partitions_per_dimension) nelements_rhs += it;
   nelements_rhs *= rhs.cp_rank;
   assert(nelements_rhs == rhs.num_factor_matrix_elements);
@@ -198,7 +198,7 @@ void cpr_parameter_pack::set(const parameter_pack& rhs){
   this->piecewise_parameter_pack::set(rhs);
   const cpr_parameter_pack& rhs_derived = dynamic_cast<const cpr_parameter_pack&>(rhs);
   if (this->factor_matrix_elements != nullptr) delete[] this->factor_matrix_elements;
-  int nelements_rhs = 0;
+  size_t nelements_rhs = 0;
   for (const auto& it : rhs_derived.num_partitions_per_dimension) nelements_rhs += it;
   nelements_rhs *= rhs_derived.cp_rank;
   assert(nelements_rhs == rhs_derived.num_factor_matrix_elements);
@@ -213,7 +213,7 @@ void cpr_parameter_pack::write_to_file(std::ofstream& file) const{
   this->piecewise_parameter_pack::write_to_file(file);
   file << this->cp_rank << "\n";
   file << this->num_factor_matrix_elements << "\n";
-  for (int i=0; i<this->num_factor_matrix_elements; i++){
+  for (size_t i=0; i<this->num_factor_matrix_elements; i++){
     if (i>0) file << ",";
     file << this->factor_matrix_elements[i];
   } file << "\n";
@@ -226,7 +226,7 @@ void cpr_parameter_pack::read_from_file(std::ifstream& file){
   file >> this->num_factor_matrix_elements;
   if (this->factor_matrix_elements == nullptr) this->factor_matrix_elements = new double[this->num_factor_matrix_elements];
   std::string temp;
-  for (int i=0; i<this->num_factor_matrix_elements; i++){
+  for (size_t i=0; i<this->num_factor_matrix_elements; i++){
     if (i==(this->num_factor_matrix_elements-1)) getline(file,temp,'\n');
     else getline(file,temp,',');
     this->factor_matrix_elements[i] = std::stod(temp);
@@ -241,7 +241,7 @@ cprg_parameter_pack::cprg_parameter_pack() : cpr_parameter_pack(){
 }
 
 cprg_parameter_pack::cprg_parameter_pack(const cprg_parameter_pack& rhs) : cpr_parameter_pack(rhs){
-  int nelements_rhs = rhs.num_models*(2+rhs.spline_degree+rhs.cp_rank);
+  size_t nelements_rhs = rhs.num_models*(2+rhs.spline_degree+rhs.cp_rank);
   this->global_models = new double[nelements_rhs];
   std::memcpy(this->global_models,rhs.global_models,nelements_rhs*sizeof(double));
   this->spline_degree = rhs.spline_degree;
@@ -258,7 +258,7 @@ cprg_parameter_pack::~cprg_parameter_pack(){
 void cprg_parameter_pack::get(parameter_pack& rhs) const{
   this->cpr_parameter_pack::get(rhs);
   cprg_parameter_pack& rhs_derived = dynamic_cast<cprg_parameter_pack&>(rhs);
-  int nelements = this->num_models*(2+this->spline_degree+this->cp_rank);
+  size_t nelements = this->num_models*(2+this->spline_degree+this->cp_rank);
   rhs_derived.global_models = new double[nelements];
   std::memcpy(rhs_derived.global_models,this->global_models,nelements*sizeof(double));
   rhs_derived.spline_degree = this->spline_degree;
@@ -268,7 +268,7 @@ void cprg_parameter_pack::get(parameter_pack& rhs) const{
 void cprg_parameter_pack::set(const parameter_pack& rhs){
   this->cpr_parameter_pack::set(rhs);
   const cprg_parameter_pack& rhs_derived = dynamic_cast<const cprg_parameter_pack&>(rhs);
-  int nelements_rhs = rhs_derived.num_models*(2+rhs_derived.spline_degree+rhs_derived.cp_rank);
+  size_t nelements_rhs = rhs_derived.num_models*(2+rhs_derived.spline_degree+rhs_derived.cp_rank);
   this->global_models = new double[nelements_rhs];
   std::memcpy(this->global_models,rhs_derived.global_models,nelements_rhs*sizeof(double));
   this->spline_degree = rhs_derived.spline_degree;
@@ -280,8 +280,8 @@ void cprg_parameter_pack::write_to_file(std::ofstream& file) const{
   this->cpr_parameter_pack::write_to_file(file);
   file << this->spline_degree;
   file << this->num_models; 
-  int nelements = this->num_models*(2+this->spline_degree+this->cp_rank);
-  for (int i=0; i<nelements; i++){
+  size_t nelements = this->num_models*(2+this->spline_degree+this->cp_rank);
+  for (size_t i=0; i<nelements; i++){
     if (i>0) file << ",";
     file << this->global_models[i];
   } file << "\n";
@@ -292,10 +292,10 @@ void cprg_parameter_pack::read_from_file(std::ifstream& file){
   this->cpr_parameter_pack::read_from_file(file);
   file >> this->spline_degree;
   file >> this->num_models; 
-  int nelements = this->num_models*(2+this->spline_degree+this->cp_rank);
+  size_t nelements = this->num_models*(2+this->spline_degree+this->cp_rank);
   if (this->global_models == nullptr) this->global_models = new double[nelements];
   std::string temp;
-  for (int i=0; i<nelements; i++){
+  for (size_t i=0; i<nelements; i++){
     if (i==(nelements-1)) getline(file,temp,'\n');
     else getline(file,temp,',');
     this->global_models[i] = std::stod(temp);
